tests/wpt/web-platform-tests/css/css-rhythm/line-height-step-parsing-001.html

<!DOCTYPE html>
<title>CSS Rhythmic Sizing: parsing line-height-step</title>
<link rel="author" title="Koji Ishii" href="mailto:kojiishi@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-rhythm-1/#line-height-step">
<meta name="assert" content="This test asserts the parsing and getComputedStyle works correctly for the line-height-step property.">
<meta name="flags" content="dom">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div title="initial value should be '0px'" data-expected="0px"></div>

<div style="line-height-step: 20px" data-expected="20px">
  <div title="line-height-step should inherit" data-expected="20px"></div>
  <div style="line-height-step: initial" data-expected="0px"></div>

  <div style="line-height-step: 40px" data-expected="40px"></div>

  <div title="'0' should be a valid length" style="line-height-step: 0" data-expected="0px"></div>

  <div title="Interger should be invalid" style="line-height-step: 1" data-expected="20px"></div>
  <div title="Negative length should be invalid" style="line-height-step: -1px" data-expected="20px"></div>
</div>

<div id="pt" style="line-height-step: 40pt"></div>

<script>
Array.prototype.forEach.call(document.querySelectorAll("[data-expected]"), function (element) {
  var expected = element.dataset.expected;
  test(function () {
    var actual = getComputedStyle(element).lineHeightStep;
    assert_equals(actual, expected);
  }, element.title || "'" + element.getAttribute("style") + "' should compute to '" + expected + "'");
});

test(function () {
  assert_equals(getComputedStyle(document.getElementById("pt")).lineHeightStep.slice(-2), "px");
}, "Computed value should be the absolute length");
</script>
